requirements.txt
yfinance
duckdb
pandas
snowflake-connector-python
snowflake-sqlalchemy
dbt-snowflake
python-dotenv
fredapi
numpy
scipy
matplotlib
pyarrow
